+++ pastell-core/NotificationMail.class.php
@@ -1,25 +1,25 @@
 <?php
 
+use Pastell\Mailer\Mailer;
+use Symfony\Bridge\Twig\Mime\TemplatedEmail;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
+
 class NotificationMail
 {
-    private $sqlQuery;
-    private $zenMail;
-    private $journal;
-    private $notification;
-    private $notificationDigestSQL;
-
-    public function __construct(Notification $notification, ZenMail $zenMail, Journal $journal, NotificationDigestSQL $notificationDigestSQL)
-    {
-        $this->journal = $journal;
-        $this->zenMail = $zenMail;
-        $this->notification = $notification;
-        $this->notificationDigestSQL = $notificationDigestSQL;
+    public function __construct(
+        private readonly Notification $notification,
+        private readonly Mailer $mailer,
+        private readonly Journal $journal,
+        private NotificationDigestSQL $notificationDigestSQL
+    ) {
     }
 
-    public function notify($id_e, $id_d, $action, $type, $message, array $attachment = [])
+    /**
+     * @throws TransportExceptionInterface
+     */
+    public function notify($id_e, $id_d, $action, $type, $message, array $attachment = []): void
     {
         $lesEmails = $this->notification->getAllInfo($id_e, $type, $action);
-
         foreach ($lesEmails as $mail_info) {
             if ($mail_info['daily_digest']) {
                 $this->register($mail_info['email'], $id_e, $id_d, $action, $type, $message);
@@ -29,36 +29,61 @@ class NotificationMail
         }
     }
 
-    private function sendMail($mail, $id_e, $id_d, $action, $type, $message, array $attachment = [])
+    /**
+     * @throws TransportExceptionInterface
+     */
+    private function sendMail($mail, $id_e, $id_d, $action, $type, $message, array $attachment = []): void
     {
-        $this->zenMail->setEmetteur("Pastell", PLATEFORME_MAIL);
-        $this->zenMail->setDestinataire($mail);
-        $this->zenMail->setSujet("[Pastell] Notification");
-        $this->zenMail->resetAttachment();
+        $url = sprintf('%s/Document/detail?id_d=%s&id_e=%d', SITE_BASE, $id_d, $id_e);
+        $templatedEmail = (new TemplatedEmail())
+            ->to($mail)
+            ->subject('[Pastell] Notification')
+            ->htmlTemplate('notification.html.twig')
+            ->context([
+                'message' => $message,
+                'url' => $url,
+                'action' => $action,
+                'type' => $type,
+                'SITE_BASE' => SITE_BASE
+            ]);
         foreach ($attachment as $filename => $filepath) {
-            $this->zenMail->addAttachment($filename, $filepath);
+            $templatedEmail->attachFromPath($filepath, $filename);
         }
-        $info = ['message' => $message,'id_e' => $id_e,'id_d' => $id_d,'action' => $action,'type' => $type];
-        $this->zenMail->setContenu(PASTELL_PATH . "/mail/notification.php", $info);
-        $this->zenMail->send();
-        $this->journal->addActionAutomatique(Journal::NOTIFICATION, $id_e, $id_d, $action, "notification envoyée à $mail");
+        $this->mailer->send($templatedEmail);
+        $this->journal->addActionAutomatique(
+            Journal::NOTIFICATION,
+            $id_e,
+            $id_d,
+            $action,
+            "notification envoyée à $mail"
+        );
     }
 
-    private function register($mail, $id_e, $id_d, $action, $type, $message)
+    private function register($mail, $id_e, $id_d, $action, $type, $message): void
     {
         $this->notificationDigestSQL->add($mail, $id_e, $id_d, $action, $type, $message);
     }
 
-    public function sendDailyDigest()
+    /**
+     * @throws TransportExceptionInterface
+     */
+    public function sendDailyDigest(): void
     {
-        $this->zenMail->setEmetteur("Pastell", PLATEFORME_MAIL);
-        $this->zenMail->setSujet("[Pastell] Notification (résumé journalier)");
         $all = $this->notificationDigestSQL->getAll();
         foreach ($all as $email => $all_info) {
-            $this->zenMail->setDestinataire($email);
-            $this->zenMail->setContenu(PASTELL_PATH . "/mail/notification-daily-digest.php", $all_info);
-            $this->zenMail->send();
-            $this->journal->addActionAutomatique(Journal::NOTIFICATION, 0, 0, false, "Résumé des notifications envoyée à $email");
+            $templatedEmail = (new TemplatedEmail())
+                ->to($email)
+                ->subject('[Pastell] Notification (résumé journalier)')
+                ->htmlTemplate('notification-daily-digest.html.twig')
+                ->context(['info' => $all_info, 'SITE_BASE' => SITE_BASE]);
+            $this->mailer->send($templatedEmail);
+            $this->journal->addActionAutomatique(
+                Journal::NOTIFICATION,
+                0,
+                0,
+                false,
+                "Résumé des notifications envoyée à $email"
+            );
             foreach ($all_info as $info) {
                 $this->notificationDigestSQL->delete($info['id_nd']);
             }